Tools - extract classes from methods in CodeAnalysisTasks
