+++ src/Tool/PhpMetrics.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Edge\QA\Tool;
+
+class PhpMetrics extends Tool
+{
+    public function __invoke()
+    {
+        $analyzedDirs = $this->options->getAnalyzedDirs();
+        $analyzedDir = reset($analyzedDirs);
+        if (count($analyzedDirs) > 1) {
+            $this->say("<error>phpmetrics analyzes only first directory {$analyzedDir}</error>");
+        }
+        $args = array(
+            $analyzedDir,
+            $this->options->ignore->phpmetrics(),
+            'extensions' => $this->config->csv('extensions')
+        );
+        if ($this->options->isSavedToFiles) {
+            $this->tool->htmlReport = $this->options->rawFile('phpmetrics.html');
+            $args['offline'] = '';
+            $args['report-html'] = \Edge\QA\escapePath($this->tool->htmlReport);
+            $args['report-xml'] = $this->options->toFile('phpmetrics.xml');
+            $configFile = $this->config->value('phpmetrics.config');
+            if ($configFile) {
+                $args['config'] = $configFile;
+            }
+        } else {
+            $args['report-cli'] = '';
+        }
+        return $args;
+    }
+}
