+++ src/Tool/Phpcpd.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Edge\QA\Tool;
+
+class Phpcpd extends Tool
+{
+    public function __invoke()
+    {
+        $args = array(
+            'progress' => '',
+            $this->options->ignore->bergmann(),
+            $this->options->getAnalyzedDirs(' '),
+            'min-lines' => $this->config->value('phpcpd.minLines'),
+            'min-tokens' => $this->config->value('phpcpd.minTokens'),
+        );
+        if ($this->options->isSavedToFiles) {
+            $args['log-pmd'] = $this->tool->getEscapedXmlFile();
+        }
+        return $args;
+    }
+}
